[SCHOL-65] Skip PDF generation and reorganize
