+++ .github/workflows/etl-pipeline-tests.yaml
@@ -63,37 +63,37 @@ jobs:
         run: |
           make unit
 
-      - name: Build Docker images with cache
-        working-directory: ./etl-pipeline
-        run: |
-          docker compose build --progress plain \
-            --build-arg BUILDKIT_INLINE_CACHE=1
+      # - name: Build Docker images with cache
+      #   working-directory: ./etl-pipeline
+      #   run: |
+      #     docker compose build --progress plain \
+      #       --build-arg BUILDKIT_INLINE_CACHE=1
 
-      - name: Start services
-        working-directory: ./etl-pipeline
-        run: |
-          docker compose up -d
+      # - name: Start services
+      #   working-directory: ./etl-pipeline
+      #   run: |
+      #     docker compose up -d
 
-      - name: Wait for containers
-        run: |
-          echo "Waiting for services..." && sleep 90
+      # - name: Wait for containers
+      #   run: |
+      #     echo "Waiting for services..." && sleep 90
 
-      - name: Run functional tests
-        env:
-          ENVIRONMENT: local
-        working-directory: ./etl-pipeline
-        run: |
-          make functional
+      # - name: Run functional tests
+      #   env:
+      #     ENVIRONMENT: local
+      #   working-directory: ./etl-pipeline
+      #   run: |
+      #     make functional
 
-      - name: Run integration tests
-        env:
-          ENVIRONMENT: local
-        working-directory: ./etl-pipeline
-        run: |
-          make integration
+      # - name: Run integration tests
+      #   env:
+      #     ENVIRONMENT: local
+      #   working-directory: ./etl-pipeline
+      #   run: |
+      #     make integration
 
-      - name: Cleanup
-        if: always()
-        working-directory: ./etl-pipeline
-        run: |
-          docker compose down
+      # - name: Cleanup
+      #   if: always()
+      #   working-directory: ./etl-pipeline
+      #   run: |
+      #     docker compose down
