+++ etl-pipeline/processes/__init__.py
@@ -13,5 +13,5 @@ from .record_clusterer import RecordClusterer
 from .record_pipeline import RecordPipelineProcess
 from .local_development.seed_local_data import SeedLocalDataProcess
 from .grin.conversion import GRINConversion
-from .grin_ingest_process import GRINIngestProcess
+from .grin.ingest import GRINIngestProcess
 from .research_assistant_chatbot import ResearchAssistantChatBot
